fix: satisfy debug handler one-per-file after app_debug extraction (#4117)
Rename shared diplomacy/flip resolve helpers outside the handler filename
prefix and split dual-handler modules so repo.debug_handler_one_per_file
passes again.

Refs #4117
