+++ packages/colonizethis_app_debug/lib/src/app_event_handler_debug_set_diplomacy_ftp.dart
@@ -1,7 +1,7 @@
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import 'app_event_handler_debug_set_diplomacy_common.dart';
+import 'debug_set_diplomacy_common.dart';
 
 DebugDiplomacyActionOutcome applyDebugDiplomacyFtp(
   Game game,
@@ -46,33 +46,6 @@ DebugDiplomacyActionOutcome applyDebugDiplomacyFtp(
   return (
     game: nextGame,
     message: 'FTP set between $factionA and $factionB.',
-    error: null,
-  );
-}
-
-DebugDiplomacyActionOutcome applyDebugDiplomacyNoFtp(
-  Game game,
-  String factionA,
-  String factionB,
-  int turn,
-) {
-  if (!hasFtpPartnership(game, factionA, factionB)) {
-    return (
-      game: game,
-      message: 'No FTP to remove between $factionA and $factionB (no change).',
-      error: null,
-    );
-  }
-  final nextFtpKeys = Set<String>.from(game.ftpPartnershipKeys)
-    ..remove(pairKey(factionA, factionB));
-  final nextGame = appendDebugDiplomacyEvents(
-    game.copyWith(ftpPartnershipKeys: nextFtpKeys),
-    turn,
-    [(type: DiplomaticEventType.ftpBroken, from: factionA, to: factionB)],
-  );
-  return (
-    game: nextGame,
-    message: 'FTP removed between $factionA and $factionB.',
     error: null,
   );
 }
