+++ packages/colonizethis_app_debug/lib/src/app_event_handler_debug_set_diplomacy_peace.dart
@@ -3,27 +3,19 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 
 import 'debug_set_diplomacy_common.dart';
 
-DebugDiplomacyActionOutcome applyDebugDiplomacyAlliance(
+DebugDiplomacyActionOutcome applyDebugDiplomacyPeace(
   Game game,
   String factionA,
   String factionB,
   int turn,
 ) {
-  if (!isGreatPower(game, factionA) || !isGreatPower(game, factionB)) {
-    return (
-      game: null,
-      message: null,
-      error: '$kDebugSetDiplomacyPrefix rejected: alliance requires both factions to be '
-          'Great Powers.',
-    );
-  }
   final rel = getRelation(game, factionA, factionB);
-  if (rel?.atWar ?? false) {
+  if (rel == null || rel.atPeace) {
     return (
       game: null,
       message: null,
-      error: '$kDebugSetDiplomacyPrefix rejected: cannot form an alliance while $factionA and '
-          '$factionB are at war.',
+      error:
+          '$kDebugSetDiplomacyPrefix rejected: $factionA and $factionB are already at peace.',
     );
   }
   final nextRelations = upsertRelation(
@@ -32,16 +24,16 @@ DebugDiplomacyActionOutcome applyDebugDiplomacyAlliance(
     factionB,
     (existing) =>
         (existing ?? DiplomacyRelation(factionId1: factionA, factionId2: factionB))
-            .copyWith(formalAlliance: true, sinceTurn: turn),
+            .copyWith(state: RelationState.atPeace, sinceTurn: turn),
   );
   final nextGame = appendDebugDiplomacyEvents(
     game.copyWith(diplomacyRelations: nextRelations),
     turn,
-    [(type: DiplomaticEventType.allianceFormed, from: factionA, to: factionB)],
+    [(type: DiplomaticEventType.peace, from: factionA, to: factionB)],
   );
   return (
     game: nextGame,
-    message: 'Formal alliance set between $factionA and $factionB.',
+    message: 'Peace set between $factionA and $factionB.',
     error: null,
   );
 }
