+++ packages/colonizethis_app_debug/lib/src/app_event_handler_debug_set_diplomacy_war_peace.dart
@@ -1,7 +1,7 @@
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import 'app_event_handler_debug_set_diplomacy_common.dart';
+import 'debug_set_diplomacy_common.dart';
 
 DebugDiplomacyActionOutcome applyDebugDiplomacyWar(
   Game game,
@@ -71,41 +71,6 @@ DebugDiplomacyActionOutcome applyDebugDiplomacyWar(
     message: 'War set between $factionA and $factionB '
         '(cleared ${clearedOvertures.length} overture(s)'
         '${ftpWasPresent ? ', removed FTP' : ''}).',
-    error: null,
-  );
-}
-
-DebugDiplomacyActionOutcome applyDebugDiplomacyPeace(
-  Game game,
-  String factionA,
-  String factionB,
-  int turn,
-) {
-  final rel = getRelation(game, factionA, factionB);
-  if (rel == null || rel.atPeace) {
-    return (
-      game: null,
-      message: null,
-      error:
-          '$kDebugSetDiplomacyPrefix rejected: $factionA and $factionB are already at peace.',
-    );
-  }
-  final nextRelations = upsertRelation(
-    game.diplomacyRelations,
-    factionA,
-    factionB,
-    (existing) =>
-        (existing ?? DiplomacyRelation(factionId1: factionA, factionId2: factionB))
-            .copyWith(state: RelationState.atPeace, sinceTurn: turn),
-  );
-  final nextGame = appendDebugDiplomacyEvents(
-    game.copyWith(diplomacyRelations: nextRelations),
-    turn,
-    [(type: DiplomaticEventType.peace, from: factionA, to: factionB)],
-  );
-  return (
-    game: nextGame,
-    message: 'Peace set between $factionA and $factionB.',
     error: null,
   );
 }
