+++ packages/colonizethis_app_debug/lib/src/debug_set_diplomacy_common.dart
